+++ src/Database/Migrations/2019_03_19_084340_create_dz_email_notifications_table.php
@@ -15,26 +15,26 @@ class CreateDzEmailNotificationsTable extends Migration {
         Schema::create('dz_email_notifications', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name');
 
-            $table->unsignedInteger('notification_id')->nullable();
+            $table->unsignedBigInteger('notification_id')->nullable();
 
             $table->string('template_name')->nullable();
             $table->string('subject', '2048')->nullable();
 
-            $table->unsignedInteger('run_condition_id')->nullable();
+            $table->unsignedBigInteger('run_condition_id')->nullable();
             $table->foreign('run_condition_id')->references('id')->on('dz_system_scripts');
 
-            $table->unsignedInteger('body_id')->nullable();
+            $table->unsignedBigInteger('body_id')->nullable();
             $table->foreign('body_id')->references('id')->on('dz_system_scripts');
 
             $table->boolean('default_users', true);
             $table->boolean('active', false);
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->foreign('notification_id')->references('id')->on('dz_notification_details');
             $table->foreign('created_by')->references('id')->on($userTable);
