+++ src/Database/Migrations/2019_07_19_055402_create_dz_inapp_notifications_table.php
@@ -15,11 +15,11 @@ class CreateDzInappNotificationsTable extends Migration {
         Schema::create('dz_inapp_notifications', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name')->nullable();
 
-            $table->unsignedInteger('notification_id')->nullable();
-            $table->unsignedInteger('platform_id')->nullable();
+            $table->unsignedBigInteger('notification_id')->nullable();
+            $table->unsignedBigInteger('platform_id')->nullable();
 
             $table->string('content', 1024)->nullable();
             $table->text('description')->nullable();
@@ -27,13 +27,13 @@ class CreateDzInappNotificationsTable extends Migration {
             $table->string('deep_link_url', 1024)->nullable();
             $table->integer('offset_end_time')->nullable();
 
-            $table->unsignedInteger('run_condition_id')->nullable();
+            $table->unsignedBigInteger('run_condition_id')->nullable();
 
             $table->boolean('default_users')->default(true);
             $table->boolean('active')->default(true);
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->foreign('notification_id')->references('id')->on('dz_notification_details');
             $table->foreign('platform_id')->references('id')->on('dz_lookup_values');
